.github/scripts/requirements.txt
# For update-artifacthub-changes.py (used in semantic-release prepare step)
PyYAML>=6.0
